move supply stats to campaigns data
